renamed and reorganized code to be more consistent with other collectors
