+++ collectors/scm/gitlab/pom.xml
@@ -1,6 +1,6 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<artifactId>gitlab-git-collector</artifactId>
+	<artifactId>gitlab-scm-collector</artifactId>
 	<packaging>jar</packaging>
 	<name>${project.groupId}:${project.artifactId}</name>
 	<description>Gitlab Collector Microservice collecting stats from Git</description>
@@ -37,7 +37,7 @@
 				<artifactId>docker-maven-plugin</artifactId>
 				<configuration>
 					<skipDockerBuild>false</skipDockerBuild>
-					<imageName>gitlab-git-collector</imageName>
+					<imageName>hygieia-gitlab-scm-collector</imageName>
 					<dockerDirectory>${project.basedir}/docker</dockerDirectory>
 					<resources>
 						<resource>
