+++ collectors/scm/gitlab/src/main/java/com/capitalone/dashboard/collector/GitlabSettings.java
@@ -1,4 +1,4 @@
-package com.capitalone.dashboard.collecteur;
+package com.capitalone.dashboard.collector;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
