+++ collectors/scm/gitlab/src/main/java/com/capitalone/dashboard/gitlab/GitlabUrlUtility.java
@@ -15,7 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.util.UriComponentsBuilder;
 
-import com.capitalone.dashboard.collecteur.GitlabSettings;
+import com.capitalone.dashboard.collector.GitlabSettings;
 import com.capitalone.dashboard.model.GitlabGitRepo;
 
 @Component
