+++ collectors/scm/gitlab/src/test/java/com/capitalone/dashboard/collector/GitlabCommitsResponseMapperTest.java
@@ -1,4 +1,4 @@
-package com.capitalone.dashboard.collecteur;
+package com.capitalone.dashboard.collector;
 
 import static org.junit.Assert.assertEquals;
