+++ collectors/scm/gitlab/src/test/java/com/capitalone/dashboard/collector/GitlabGitCollectorTaskTest.java
@@ -1,4 +1,4 @@
-package com.capitalone.dashboard.collecteur;
+package com.capitalone.dashboard.collector;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertSame;
@@ -25,6 +25,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.ResourceAccessException;
 
+import com.capitalone.dashboard.collector.GitlabGitCollectorTask;
+import com.capitalone.dashboard.collector.GitlabSettings;
 import com.capitalone.dashboard.gitlab.DefaultGitlabGitClient;
 import com.capitalone.dashboard.model.Collector;
 import com.capitalone.dashboard.model.CollectorType;
